chore(zero-cache): re-enable single process mode
