+++ apps/zbugs/package.json
@@ -13,7 +13,7 @@
     "lint": "eslint --ext .ts,.tsx,.js,.jsx src/",
     "preview": "vite preview",
     "zero": "npm run zero-build-schema && tsx ../../packages/zero-cache/src/server/multi/main.ts",
-    "zero-brk": "export SINGLE_PROCESS=1 && npm run zero-build-schema && tsx --inspect-brk ../../packages/zero-cache/src/server/multi/main.ts",
+    "zero-brk": "npm run zero-build-schema && tsx --inspect-brk ../../packages/zero-cache/src/server/debug/single.ts",
     "transform-query": "npm run zero-build-schema && tsx ../../packages/zero-cache/src/scripts/transform-query.ts",
     "run-query": "npm run zero-build-schema && tsx ../../packages/zero-cache/src/scripts/run-query.ts",
     "run-query-brk": "npm run zero-build-schema && tsx --inspect-brk ../../packages/zero-cache/src/scripts/run-query.ts",
