+++ packages/zero-cache/src/server/debug/single.ts
@@ -1,0 +1,4 @@
+import {setSingleProcessMode} from '../../types/processes.ts';
+import {runWorker} from '../multi/run-worker.ts';
+setSingleProcessMode(true);
+await runWorker(null, process.env);
